chore: 기존 모놀리즈 레포 코드들을 app 모듈로 이동
